Improve performance of sudoku solver by smarter pruning
Prefer to explore more constrained spaces first. This massively reduces the
number of paths that need to be explored because you are more likely to make the
right choice early on.

Consider the extreme case, where there is an unfilled position with only one
possible choice. Filling it early will trim all subsequent choices, essentially
for free.